pyproject.toml
[project]
name = "nuke-fm"
version = "0.1.0"
description = "Read-only prediction market catalog for nuke.fm MVP"
authors = [
    { name = "distbit", email = "adp01@pm.me" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nukefm"]
